# Use a much faster hash for TypeId