TypeId is already a fully-hashed u64, so there is no need to hash it further.
Instead, just use nohash-hasher to pass it through unchanged.