@mjhostet 2020-10-18
This general trick is a measurable win in a compiler I work on, but I wasn't sure how to benchmark here. I tried "cargo bench" and it just prints a bunch of "ignored" lines but I didn't look into it.