@Ralith 2020-10-18
> I tried "cargo bench" and it just prints a bunch of "ignored" lines but I didn't look into it.

`cargo bench --all-features` is needed as the benchmarks use the macros. They're not *super* comprehensive since I've just been filling them out as needed to answer specific performance questions.