@cart 2020-10-18
Something to point out: this might not work forever as rustc devs are considering making TypeId a u128 instead of a u64 internally (which nohash doesn't support):

https://github.com/rust-lang/rust/pull/75923