@Ralith 2020-10-18
Good catch, thanks! We should probably roll our own which tolerates arbitrary-length inputs, then. Presuming the statistical properties are similar we should be able to just truncate.