@mjhostet 2020-10-18
This is about 33% faster on the spawn_batch benchmark (34147 ns -> 25611 ns), but strangely 11% slower on iterate_100k (179849 ns -> 202057 ns). I don't have an explanation.