@Ralith 2020-10-20
Splitting the change-of-archetype case of `QueryIter::next` into its own function narrows the gap some with this change, but makes things worse on master, bringing both to ~250μs.