@cart 2020-10-20
Yeah given that iteration speed is the "strong point" of an archetypal ecs, I think we should optimize for that over spawn speed.