@Imberflur 2020-10-20
On my system (i5-4690) I see
spawn_batch: ~45us-> ~32us
iterate_100k: ~207us -> ~208us