@Ralith 2020-10-20
The outlining change I described above makes this change faster than master on today's nightly (but not on rustc 1.47), even though it still makes master slower:

```rust
diff --git a/src/query.rs b/src/query.rs
index e635e33..8c9d80f 100644
--- a/src/query.rs
+++ b/src/query.rs
@@ -432,6 +432,21 @@ pub struct QueryIter<'q, 'w, Q: Query> {
     iter: Option<ChunkIter<Q>>,
 }
 
+impl<'q, 'w, Q: Query> QueryIter<'q, 'w, Q> {
+    fn next_archetype(&mut self) -> Option<()> {
+        let archetype = self.borrow.archetypes.get(self.archetype_index as usize)?;
+        self.archetype_index += 1;
+        unsafe {
+            self.iter = Q::Fetch::get(archetype, 0).map(|fetch| ChunkIter {
+                entities: archetype.entities(),
+                fetch,
+                len: archetype.len(),
+            });
+        }
+        Some(())
+    }
+}
+
 unsafe impl<'q, 'w, Q: Query> Send for QueryIter<'q, 'w, Q> {}
 unsafe impl<'q, 'w, Q: Query> Sync for QueryIter<'q, 'w, Q> {}
 
@@ -442,17 +457,7 @@ impl<'q, 'w, Q: Query> Iterator for QueryIter<'q, 'w, Q> {
     fn next(&mut self) -> Option<Self::Item> {
         loop {
             match self.iter {
-                None => {
-                    let archetype = self.borrow.archetypes.get(self.archetype_index as usize)?;
-                    self.archetype_index += 1;
-                    unsafe {
-                        self.iter = Q::Fetch::get(archetype, 0).map(|fetch| ChunkIter {
-                            entities: archetype.entities(),
-                            fetch,
-                            len: archetype.len(),
-                        });
-                    }
-                }
+                None => self.next_archetype()?,
                 Some(ref mut iter) => match unsafe { iter.next() } {
                     None => {
                         self.iter = None;
```

I feel like something about `QueryIter::next` is really throwing llvm for a loop. Maybe @cart's inlining changes would have an impact here?