@Ralith 2020-10-20
On advice from @FeepingCreature I got rid of the outer branch in `QueryIter::next`, ~doubling performance and annihilating the confusing discrepancies here. See https://github.com/Ralith/hecs/pull/95.