@cart 2020-10-22
Just ported #95 and #96 to bevy_hecs. As I sort of suspected, it had zero effect (maybe a slighty negative effect) on simple iteration performance, buuuut it did give some nice boosts to other benches like "fragmented iteration":

![image](https://user-images.githubusercontent.com/2694663/96805000-6772e480-13c5-11eb-821b-9c5d4ab293fe.png)

For reference: on my machine hecs went from ~1800ns to 990ns on simple_iter. I assume the 200ns discrepancy between hecs and bevy is the overhead of boxed systems / query setup / Fetch::should_skip.

I'm gonna go ahead and call this a huge win for hecs and a medium-sized win for bevy :)