@cart 2020-10-22
Ah wait we aren't using bevy's System abstraction in those benches. So thats not where the overhead comes from.