@cart 2020-10-22
Ooh I just ported the TypeId hasher and thats also a huge win for bevy:

![image](https://user-images.githubusercontent.com/2694663/96805873-7c507780-13c7-11eb-8511-ac0fdb9f100c.png)