@Ralith 2020-10-22
Interesting! Can you try running hecs' benchmarks before/after #100? If you see no change there as well I'll write it off to a quirk of my environment (??) and keep the change.