@Ralith 2020-10-22
> Fetch::should_skip

Provided there's sufficient inlining (which should be achievable by annotations if it's not happening on its own) I think this should be zero-cost in the sense that queries which don't use it will have the branches compiled out.