@cart 2020-10-22
This is the result of running the benches on the "revert" branch first, then running them again on master, so the "improvements" are _with_ "Use a much faster hash for TypeId"

![image](https://user-images.githubusercontent.com/2694663/96807566-c176a880-13cb-11eb-9515-92ca9cc56bbf.png)


Based on these results, I don't think reverting is the right call (at least for these benchmarks)